chore: remove unused const
