+++ cmd/gossamer/config.go
@@ -49,8 +49,6 @@ var (
 	kusamaName   = "kusama"
 	polkadotName = "polkadot"
 	devName      = "dev"
-
-	storagePath = "storage"
 )
 
 // loadConfigFile loads a default config file if --chain is specified, a specific
@@ -789,7 +787,7 @@ func updateDotConfigFromGenesisJSONRaw(tomlCfg ctoml.Config, cfg *dot.Config) {
 // updateDotConfigFromGenesisData updates the configuration from genesis data of an initialised node
 func updateDotConfigFromGenesisData(ctx *cli.Context, cfg *dot.Config) error {
 	// initialise database using data directory
-	db, err := state.SetupDatabase(cfg.Global.BasePath, false)
+	db, err := utils.SetupDatabase(cfg.Global.BasePath, false)
 	if err != nil {
 		return fmt.Errorf("failed to create database: %s", err)
 	}
