+++ dot/state/base.go
@@ -20,23 +20,12 @@ import (
 	"encoding/binary"
 	"encoding/json"
 	"fmt"
-	"path/filepath"
 
 	"github.com/ChainSafe/gossamer/lib/common"
 	"github.com/ChainSafe/gossamer/lib/genesis"
 
 	"github.com/ChainSafe/chaindb"
 )
-
-const defaultDatabaseDir = "storage"
-
-// SetupDatabase will return an instance of database based on basepath
-func SetupDatabase(basepath string, inMemory bool) (chaindb.Database, error) {
-	return chaindb.NewBadgerDB(&chaindb.Config{
-		DataDir:  filepath.Join(basepath, defaultDatabaseDir),
-		InMemory: inMemory,
-	})
-}
 
 // BaseState is a wrapper for the chaindb.Database, without any prefixes
 type BaseState struct {
