+++ dot/state/service.go
@@ -26,6 +26,7 @@ import (
 	"github.com/ChainSafe/gossamer/dot/types"
 	"github.com/ChainSafe/gossamer/lib/blocktree"
 	"github.com/ChainSafe/gossamer/lib/trie"
+	"github.com/ChainSafe/gossamer/lib/utils"
 
 	"github.com/ChainSafe/chaindb"
 	log "github.com/ChainSafe/log15"
@@ -96,7 +97,7 @@ func (s *Service) Start() error {
 		}
 
 		// initialise database
-		db, err = SetupDatabase(basepath, false)
+		db, err = utils.SetupDatabase(basepath, false)
 		if err != nil {
 			return err
 		}
@@ -296,7 +297,7 @@ func (s *Service) Stop() error {
 func (s *Service) Import(header *types.Header, t *trie.Trie, firstSlot uint64) error {
 	var err error
 	// initialise database using data directory
-	s.db, err = SetupDatabase(s.dbPath, s.isMemDB)
+	s.db, err = utils.SetupDatabase(s.dbPath, s.isMemDB)
 	if err != nil {
 		return fmt.Errorf("failed to create database: %s", err)
 	}
